refactor(amp-pwa): replaced request.mode
Replaced service worker route that relied on request.mode. There are now
three routes: one that matches all non-amp resources and responds with
cacheFirst, one that matches the amp pages and responds with the shell,
and one that checks that the url ends with shell=false and responds with
cacheFirst. Updated the shell JS to add the shell=false parameter to the
request for the AMP pages.

Change-Id: I17b81a5d5fc6a26edf1b0ad516d060c401b4ac6e