feat(*): update to Workbox v3 alpha (not final)
This commit updates the starter & solution code to use
Workbox v3 alpha 5.

NOTE: The solution code is currently using local copies
of the Workbox libraries since the v3 alpha CDN is not
always up to date. These will need to be removed in favor
of the canonical CDN library. The starter & solution code
will also of course need to be re-tested once v3 is out
of alpha, or if any significant changes are made.

Code at this commit stage is only for testing.